Create Flood Warning and Evacuation System
Problem Statement: There is a lack of reliable and timely warning systems in flood-prone areas, which can result in loss of life and damage to property. An effective and affordable solution is needed to detect rising water levels and notify people in the area to evacuate in a timely manner.